requirements.txt
asgiref==3.8.1
colorama==0.4.6
distlib==0.3.9
Django==5.1.6
djangorestframework==3.15.2
filelock==3.17.0
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
virtualenv==20.29.2
Werkzeug==3.1.3
django-environ
django-extensions
stripe==12.1.0
